+++ capa/ida/plugin/form.py
@@ -776,13 +776,15 @@ class CapaExplorerForm(idaapi.PluginForm):
 
                 try:
                     meta = capa.ida.helpers.collect_metadata([Path(settings.user[CAPA_SETTINGS_RULE_PATH])])
-                    capabilities, counts = capa.capabilities.common.find_capabilities(
+                    capabilities = capa.capabilities.common.find_capabilities(
                         ruleset, self.feature_extractor, disable_progress=True
                     )
 
-                    meta.analysis.feature_counts = counts["feature_counts"]
-                    meta.analysis.library_functions = counts["library_functions"]
-                    meta.analysis.layout = capa.loader.compute_layout(ruleset, self.feature_extractor, capabilities)
+                    meta.analysis.feature_counts = capabilities.feature_counts
+                    meta.analysis.library_functions = capabilities.library_functions
+                    meta.analysis.layout = capa.loader.compute_layout(
+                        ruleset, self.feature_extractor, capabilities.matches
+                    )
                 except UserCancelledError:
                     logger.info("User cancelled analysis.")
                     return False
@@ -832,7 +834,7 @@ class CapaExplorerForm(idaapi.PluginForm):
 
                 try:
                     self.resdoc_cache = capa.render.result_document.ResultDocument.from_capa(
-                        meta, ruleset, capabilities
+                        meta, ruleset, capabilities.matches
                     )
                 except Exception as e:
                     logger.exception("Failed to collect results (error: %s)", e)