+++ CHANGELOG.md
@@ -3,15 +3,21 @@
 ## master (unreleased)
 
 ### New Features
+
 - add warning for dynamic .NET samples #1864 @v1bh475u
+- add span-of-calls scope to match features against a across a sliding window of API calls within a thread @williballenthin #2532
 
 ### Breaking Changes
+
 - remove `is_static_limitation` class function from `capa.rules.Rule` to inline function in `capa.capabilities.common.py`
- 
-### New Rules (2)
+- add span-of-calls scope to rule format
+- capabilities functions return dataclasses instead of tuples
+
+### New Rules (3)
 
 - data-manipulation/encryption/rsa/encrypt-data-using-rsa-via-embedded-library @Ana06
 - data-manipulation/encryption/use-bigint-function @Ana06
+- nursery/dynamic-add-veh wballenthin@google.com
 -
 
 ### Bug Fixes